Add support for Python API v2

+++ simvue_integrations/connectors/openfoam.py
@@ -95,7 +95,8 @@ class OpenfoamRun(WrappedRun):
         if self.upload_as_zip:
             zip_file.close()
             self.save_file(out_zip, file_type)
-            pathlib.Path(out_zip).unlink()
+            if not self._sv_obj._offline:
+                pathlib.Path(out_zip).unlink()
 
     @mp_tail_parser.log_parser
     def _log_parser(
@@ -124,7 +125,7 @@ class OpenfoamRun(WrappedRun):
         )
         exp2: re.Pattern[str] = re.compile(r"^ExecutionTime = ([0-9.]+) s")
         metrics = {}
-        header_metadata = {}
+        header_metadata = {"openfoam": {}}
         title = False
         header = False
         solver_info = False
@@ -177,7 +178,7 @@ class OpenfoamRun(WrappedRun):
                 if key == "exec":
                     current_process = value
                 else:
-                    header_metadata[f"openfoam.{key}"] = value
+                    header_metadata["openfoam"][key] = value
 
             # Log events for any initial solver info
             if solver_info and line:
